rafs: move overlay related code into builder/core/overlay.rs
Move overlay related code into builder/core/overlay.rs, for better
maintenance.
