+++ rafs/src/builder/core/node.rs
@@ -16,7 +16,6 @@ use std::os::linux::fs::MetadataExt;
 use std::os::macos::fs::MetadataExt;
 use std::os::unix::ffi::OsStrExt;
 use std::path::{Component, Path, PathBuf};
-use std::str::FromStr;
 use std::sync::Arc;
 
 use anyhow::{anyhow, bail, ensure, Context, Error, Result};
@@ -31,6 +30,7 @@ use sha2::digest::Digest;
 
 use super::chunk_dict::{ChunkDict, DigestWithBlobIndex};
 use super::context::{ArtifactWriter, BlobContext, BlobManager, BootstrapContext, BuildContext};
+use super::overlay::Overlay;
 use super::tree::Tree;
 use crate::metadata::chunk::ChunkWrapper;
 use crate::metadata::inode::{new_v6_inode, InodeWrapper};
@@ -56,112 +56,6 @@ const RAFS_V5_VIRTUAL_ENTRY_SIZE: u64 = 8;
 
 /// Filesystem root path for Unix OSs.
 pub const ROOT_PATH_NAME: &[u8] = &[b'/'];
-
-/// Prefix for OCI whiteout file.
-pub const OCISPEC_WHITEOUT_PREFIX: &str = ".wh.";
-/// Prefix for OCI whiteout opaque.
-pub const OCISPEC_WHITEOUT_OPAQUE: &str = ".wh..wh..opq";
-/// Extended attribute key for Overlayfs whiteout opaque.
-pub const OVERLAYFS_WHITEOUT_OPAQUE: &str = "trusted.overlay.opaque";
-
-// # Overlayfs Whiteout
-//
-// In order to support rm and rmdir without changing the lower filesystem, an overlay filesystem
-// needs to record in the upper filesystem that files have been removed. This is done using
-// whiteouts and opaque directories (non-directories are always opaque).
-//
-// A whiteout is created as a character device with 0/0 device number. When a whiteout is found
-// in the upper level of a merged directory, any matching name in the lower level is ignored,
-// and the whiteout itself is also hidden.
-//
-// A directory is made opaque by setting the xattr “trusted.overlay.opaque” to “y”. Where the upper
-// filesystem contains an opaque directory, any directory in the lower filesystem with the same
-// name is ignored.
-//
-// # OCI Image Whiteout
-// - A whiteout file is an empty file with a special filename that signifies a path should be
-//   deleted.
-// - A whiteout filename consists of the prefix .wh. plus the basename of the path to be deleted.
-// - As files prefixed with .wh. are special whiteout markers, it is not possible to create a
-//   filesystem which has a file or directory with a name beginning with .wh..
-// - Once a whiteout is applied, the whiteout itself MUST also be hidden.
-// - Whiteout files MUST only apply to resources in lower/parent layers.
-// - Files that are present in the same layer as a whiteout file can only be hidden by whiteout
-//   files in subsequent layers.
-// - In addition to expressing that a single entry should be removed from a lower layer, layers
-//   may remove all of the children using an opaque whiteout entry.
-// - An opaque whiteout entry is a file with the name .wh..wh..opq indicating that all siblings
-//   are hidden in the lower layer.
-#[derive(Clone, Copy, Debug, PartialEq)]
-pub enum WhiteoutType {
-    OciOpaque,
-    OciRemoval,
-    OverlayFsOpaque,
-    OverlayFsRemoval,
-}
-
-impl WhiteoutType {
-    pub fn is_removal(&self) -> bool {
-        *self == WhiteoutType::OciRemoval || *self == WhiteoutType::OverlayFsRemoval
-    }
-}
-
-#[derive(Clone, Copy, PartialEq)]
-pub enum WhiteoutSpec {
-    /// https://github.com/opencontainers/image-spec/blob/master/layer.md#whiteouts
-    Oci,
-    /// "whiteouts and opaque directories" in https://www.kernel.org/doc/Documentation/filesystems/overlayfs.txt
-    Overlayfs,
-    /// No whiteout spec, which will build all `.wh.*` and `.wh..wh..opq` files into bootstrap.
-    None,
-}
-
-impl Default for WhiteoutSpec {
-    fn default() -> Self {
-        Self::Oci
-    }
-}
-
-impl FromStr for WhiteoutSpec {
-    type Err = Error;
-
-    fn from_str(s: &str) -> Result<Self> {
-        match s {
-            "oci" => Ok(Self::Oci),
-            "overlayfs" => Ok(Self::Overlayfs),
-            "none" => Ok(Self::None),
-            _ => Err(anyhow!("invalid whiteout spec")),
-        }
-    }
-}
-
-#[allow(dead_code)]
-#[derive(Clone, Debug, PartialEq)]
-pub enum Overlay {
-    Lower,
-    UpperAddition,
-    UpperOpaque,
-    UpperRemoval,
-    UpperModification,
-}
-
-impl Overlay {
-    pub fn is_lower_layer(&self) -> bool {
-        self == &Overlay::Lower
-    }
-}
-
-impl Display for Overlay {
-    fn fmt(&self, f: &mut Formatter) -> fmt::Result {
-        match self {
-            Overlay::Lower => write!(f, "LOWER"),
-            Overlay::UpperAddition => write!(f, "ADDED"),
-            Overlay::UpperOpaque => write!(f, "OPAQUED"),
-            Overlay::UpperRemoval => write!(f, "REMOVED"),
-            Overlay::UpperModification => write!(f, "MODIFIED"),
-        }
-    }
-}
 
 /// Where the chunk data is actually stored.
 #[derive(Clone, Hash, PartialEq, Eq)]
@@ -1430,87 +1324,6 @@ impl Node {
         }
 
         Ok(())
-    }
-}
-
-// OCI and Overlayfs whiteout handling.
-impl Node {
-    /// Check whether the inode is a special overlayfs whiteout file.
-    pub fn is_overlayfs_whiteout(&self, spec: WhiteoutSpec) -> bool {
-        if spec != WhiteoutSpec::Overlayfs {
-            return false;
-        }
-        self.inode.is_chrdev()
-            && nydus_utils::compact::major_dev(self.info.rdev) == 0
-            && nydus_utils::compact::minor_dev(self.info.rdev) == 0
-    }
-
-    /// Check whether the inode (directory) is a overlayfs whiteout opaque.
-    pub fn is_overlayfs_opaque(&self, spec: WhiteoutSpec) -> bool {
-        if spec != WhiteoutSpec::Overlayfs || !self.is_dir() {
-            return false;
-        }
-
-        // A directory is made opaque by setting the xattr "trusted.overlay.opaque" to "y".
-        if let Some(v) = self
-            .info
-            .xattrs
-            .get(&OsString::from(OVERLAYFS_WHITEOUT_OPAQUE))
-        {
-            if let Ok(v) = std::str::from_utf8(v.as_slice()) {
-                return v == "y";
-            }
-        }
-
-        false
-    }
-
-    /// Get whiteout type to process the inode.
-    pub fn whiteout_type(&self, spec: WhiteoutSpec) -> Option<WhiteoutType> {
-        if self.overlay == Overlay::Lower {
-            return None;
-        }
-
-        match spec {
-            WhiteoutSpec::Oci => {
-                if let Some(name) = self.name().to_str() {
-                    if name == OCISPEC_WHITEOUT_OPAQUE {
-                        return Some(WhiteoutType::OciOpaque);
-                    } else if name.starts_with(OCISPEC_WHITEOUT_PREFIX) {
-                        return Some(WhiteoutType::OciRemoval);
-                    }
-                }
-            }
-            WhiteoutSpec::Overlayfs => {
-                if self.is_overlayfs_whiteout(spec) {
-                    return Some(WhiteoutType::OverlayFsRemoval);
-                } else if self.is_overlayfs_opaque(spec) {
-                    return Some(WhiteoutType::OverlayFsOpaque);
-                }
-            }
-            WhiteoutSpec::None => {
-                return None;
-            }
-        }
-
-        None
-    }
-
-    /// Get original filename from a whiteout filename.
-    pub fn origin_name(&self, t: WhiteoutType) -> Option<&OsStr> {
-        if let Some(name) = self.name().to_str() {
-            if t == WhiteoutType::OciRemoval {
-                // the whiteout filename prefixes the basename of the path to be deleted with ".wh.".
-                return Some(OsStr::from_bytes(
-                    name[OCISPEC_WHITEOUT_PREFIX.len()..].as_bytes(),
-                ));
-            } else if t == WhiteoutType::OverlayFsRemoval {
-                // the whiteout file has the same name as the file to be deleted.
-                return Some(name.as_ref());
-            }
-        }
-
-        None
     }
 }
